Clang warns faze 2

+++ City.cpp
@@ -6,8 +6,8 @@ void City_Plane::generate(size_t width, size_t height) {
   cities_.push_back(std::make_unique<First_City>());
 
   // TO DO : Debug
-  for(int i = width / 2 - 2; i < width / 2 + 3; ++i) {
-    for(int j = height / 2 - 2; j < height / 2 + 3; ++j) {
+  for(size_t i = width / 2 - 2; i < width / 2 + 3; ++i) {
+    for(size_t j = height / 2 - 2; j < height / 2 + 3; ++j) {
       city_land_[j][i] = true;
     }
   }
@@ -38,8 +38,8 @@ void City_Plane::draw(sf::RenderWindow& window, const Player& player) {
 
   sf::FloatRect player_box = player.bounding_box();
 
-  for(int j = player.y_range().first; j < player.y_range().second; ++j) {
-    for(int i = player.x_range().first; i < player.x_range().second; ++i) {
+  for(size_t j = player.y_range().first; j < player.y_range().second; ++j) {
+    for(size_t i = player.x_range().first; i < player.x_range().second; ++i) {
       if(city_horz_walls_[j][i]) { // ignores nullptr
         if(city_horz_walls_[j][i]->is_overlapped(player_box)) {
           city_horz_walls_[j][i]->transparent_draw(window);
